feat: 文件夹获取大小、文件压缩、文件解压等抽离为异步方法，增加异常错误信息展示

+++ BKGalMgr/Common/FileSystemMisc.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.IO;
+using System.IO.Compression;
 using System.Linq;
 using System.Text;
 using System.Threading;
@@ -41,21 +42,54 @@ public class FileSystemMisc
         return await folderPicker.PickSingleFolderAsync();
     }
 
+    public static (bool success, string message) CopyFile(
+        string sourceFileName,
+        string destFileName,
+        bool overwrite = true
+    )
+    {
+        try
+        {
+            // Directory.CreateDirectory 是一个相当灵活的方法。如果目标文件夹不存在，它会自动创建；
+            // 如果目标文件夹已经存在，它会忽略这个操作。同时，它还会沿途创建所有不存在的文件夹（类似 mkdir 的 -p 参数）
+            var destParentFolderPath = Path.GetDirectoryName(destFileName);
+            if (destParentFolderPath != null)
+                Directory.CreateDirectory(destParentFolderPath);
+            File.Copy(sourceFileName, destFileName, overwrite);
+        }
+        catch (Exception e)
+        {
+            return (false, e.Message);
+        }
+        return (true, "");
+    }
+
+    public static async Task<(bool success, string message)> CopyFileAsync(
+        string sourceFileName,
+        string destFileName,
+        bool overwrite = true
+    )
+    {
+        return await Task.Run(() => CopyFile(sourceFileName, destFileName, overwrite));
+    }
+
     // https://blog.coldwind.top/posts/how-to-copy-folder/
     public static (bool success, string message) CopyDirectory(string sourceFolderPath, string targetFolderPath)
     {
+        if (!Directory.Exists(sourceFolderPath))
+            return (false, $"{sourceFolderPath} is invalid");
         try
         {
             Directory.CreateDirectory(targetFolderPath);
 
-            foreach (string filePath in Directory.GetFiles(sourceFolderPath, "*.*", SearchOption.AllDirectories))
+            foreach (string filePath in Directory.GetFiles(sourceFolderPath, "*", SearchOption.AllDirectories))
             {
                 var relativePath = Path.GetRelativePath(sourceFolderPath, filePath);
                 var targetFilePath = Path.Combine(targetFolderPath, relativePath);
-                var subTargetFolderPath = Path.GetDirectoryName(targetFilePath);
                 // Path.GetDirectoryName 方法有可能返回空。这一情况通常发生在文件位于根目录的情况（例如 Windows 的 C:\，或 Unix 的 /）。
-                if (subTargetFolderPath != null)
-                    Directory.CreateDirectory(subTargetFolderPath);
+                var destParentFolderPath = Path.GetDirectoryName(targetFilePath);
+                if (destParentFolderPath != null)
+                    Directory.CreateDirectory(destParentFolderPath);
                 File.Copy(filePath, targetFilePath, true);
             }
         }
@@ -68,6 +102,8 @@ public class FileSystemMisc
 
     public static (bool success, string message) MoveOrCopyDirectory(string sourceDirName, string destDirName)
     {
+        if (!Directory.Exists(sourceDirName))
+            return (false, $"{sourceDirName} is invalid");
         // move
         try
         {
@@ -81,7 +117,15 @@ public class FileSystemMisc
             {
                 // move need delete target folder
                 if (Directory.Exists(destDirName))
+                {
                     Directory.Delete(destDirName, true);
+                }
+                else
+                {
+                    var destParentFolderPath = Path.GetDirectoryName(destDirName);
+                    if (destParentFolderPath != null)
+                        Directory.CreateDirectory(destParentFolderPath);
+                }
                 Directory.Move(sourceDirName, destDirName);
                 return (true, "");
             }
@@ -109,6 +153,13 @@ public class FileSystemMisc
         return GetDirectorySize(new DirectoryInfo(sourceDir), recurse);
     }
 
+    public static async Task<long> GetDirectorySizeAsync(string path, bool recurse = true)
+    {
+        if (!Directory.Exists(path))
+            return 0;
+        return await Task.Run((() => GetDirectorySize(new DirectoryInfo(path), recurse)));
+    }
+
     private static long GetDirectorySize(DirectoryInfo di, bool recurse = true)
     {
         long size = 0;
@@ -119,7 +170,7 @@ public class FileSystemMisc
 
         if (recurse)
         {
-            DirectoryInfo[] diEntries = di.GetDirectories("*.*", SearchOption.TopDirectoryOnly);
+            DirectoryInfo[] diEntries = di.GetDirectories("*", SearchOption.TopDirectoryOnly);
             Parallel.For<long>(
                 0,
                 diEntries.Length,
@@ -168,5 +219,63 @@ public class FileSystemMisc
             return (true, "");
 
         return await Task.Run(() => DeleteDirectory(path));
+    }
+
+    public static async Task<(bool success, string message)> CreateZipFromDirectoryAsync(
+        string sourceDirectoryName,
+        string destinationArchiveFileName,
+        CompressionLevel compressionLevel,
+        bool includeBaseDirectory = false
+    )
+    {
+        return await Task.Run(() =>
+        {
+            try
+            {
+                if (File.Exists(destinationArchiveFileName))
+                {
+                    File.Delete(destinationArchiveFileName);
+                }
+                else
+                {
+                    var destParentFolderPath = Path.GetDirectoryName(destinationArchiveFileName);
+                    if (destParentFolderPath != null)
+                        Directory.CreateDirectory(destParentFolderPath);
+                }
+
+                ZipFile.CreateFromDirectory(
+                    sourceDirectoryName,
+                    destinationArchiveFileName,
+                    compressionLevel,
+                    includeBaseDirectory
+                );
+            }
+            catch (Exception e)
+            {
+                return (false, e.Message);
+            }
+            return (true, "");
+        });
+    }
+
+    public static async Task<(bool success, string message)> ExtractZipFromDirectoryAsync(
+        string sourceArchiveFileName,
+        string destinationDirectoryName,
+        bool overwriteFiles = true
+    )
+    {
+        return await Task.Run(() =>
+        {
+            try
+            {
+                Directory.CreateDirectory(destinationDirectoryName);
+                ZipFile.ExtractToDirectory(sourceArchiveFileName, destinationDirectoryName, overwriteFiles);
+            }
+            catch (Exception e)
+            {
+                return (false, e.Message);
+            }
+            return (true, "");
+        });
     }
 }
